+++ R/functions__lazy.R
@@ -685,7 +685,7 @@ pl$var = function(column, ddof = 1) {
 #' Concat the arrays in a Series dtype List in linear time.
 #' @description Folds the expressions from left to right, keeping the first non-null value.
 #' @name pl_concat_list
-#' @param exprs list of Expr or Series or strings or a mix, or a char vector
+#' @param exprs list of Into<Expr>, strings interpreted as column names
 #' @return Expr
 #'
 #' @keywords Expr_new
@@ -710,9 +710,8 @@ pl$var = function(column, ddof = 1) {
 #' ))$alias("alice")$lit_to_s()
 #'
 pl$concat_list = function(exprs) {
-  l_expr = lapply(as.list(exprs), wrap_e)
-  pra = do.call(construct_ProtoExprArray, l_expr)
-  unwrap(concat_list(pra))
+  concat_list(as.list(exprs)) |>
+    unwrap(" in pl$concat_list():")
 }