+++ src/rust/src/rlib.rs
@@ -50,9 +50,10 @@ fn sum_exprs(exprs: &ProtoExprArray) -> Expr {
 }
 
 #[extendr]
-fn concat_list(exprs: &ProtoExprArray) -> Result<Expr, String> {
-    let exprs = exprs.to_vec("select");
-    Ok(Expr(pl::concat_list(exprs).map_err(|err| err.to_string())?))
+fn concat_list(exprs: Robj) -> RResult<Expr> {
+    pl::concat_list(robj_to!(VecPLExprCol, exprs)?)
+        .map_err(polars_to_rpolars_err)
+        .map(Expr)
 }
 
 #[extendr]
@@ -89,11 +90,9 @@ fn r_date_range_lazy(
     }
 }
 
-//TODO py-polars have some fancy transmute conversions TOExprs trait, maybe imple that too
-//for now just use inner directly
 #[extendr]
-fn as_struct(exprs: Robj) -> Result<Expr, String> {
-    Ok(pl::as_struct(crate::utils::list_expr_to_vec_pl_expr(exprs, true, true)?).into())
+fn as_struct(exprs: Robj) -> RResult<Expr> {
+    Ok(pl::as_struct(robj_to!(VecPLExprNamed, exprs)?).into())
 }
 
 #[extendr]