+++ NEWS.md
@@ -9,6 +9,8 @@
   `pl$sum_horizontal()` instead for rowwise computation (#508).
 - `$is_not()` is deprecated and will be removed in 0.12.0. Use `$not_()` instead 
   (#511).
+- `pl$concat_list()`: elements being strings are now interpreted as column names. Use `pl$lit` to
+  concat with a string.
 
 ## What's changed